legacy code removed
